+++ src/main/java/Adbistju/system/comparator/Comparator.java
@@ -1,4 +1,4 @@
-package Adbistju.system.comarator;
+package Adbistju.system.comparator;
 
 public abstract class Comparator {
     public abstract boolean check(String name);
